+++ examples/browser/webpack.config.js
@@ -4,6 +4,6 @@ module.exports = {
   entry: './src/index.js',
   output: {
     filename: 'bundle.js',
-    path: path.resolve(__dirname, 'dist')
+    path: path.resolve(__dirname, 'lib')
   },
 };