+++ tests/Opal.test.js
@@ -2,7 +2,7 @@ describe('Opal', () => {
     it('keeps Function.prototype', () => {
         const prev = Object.keys(Function.prototype);
 
-        require('../dist/opal');
+        require('../lib/opal');
 
         expect(Object.keys(Function.prototype)).toEqual(prev);
     });