+++ CHANGELOG.md
@@ -5,6 +5,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## [Unreleased]
 ### Fixed
 - Fix destruction of Function.prototype
+### Changed
+- Rename `dist` to `lib`
 
 ## [0.2.0] - 2017-08-22
 ### Added