+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
-## [Unreleased]
+## [0.2.1] - 2017-08-22
 ### Fixed
 - Fix destruction of Function.prototype
 ### Changed